requirements.txt
django==1.8.6
django-haystack==2.4.1
django-taggit==0.18.1


